refactor(mentor): 비밀번호 토글 모듈 'util' -> 'auth' 함수로 변경 + 리팩토링
