+++ login.html
@@ -11,7 +11,7 @@
       href="images/favicon-16x16.png"
     />
     <link rel="stylesheet" href="css/pages/auth.css" />
-    <script src="script/login.js" type="module" defer></script>
+    <script src="script/auth/login.js" type="module" defer></script>
   </head>
   <body>
     <main class="login auth">
@@ -56,6 +56,7 @@
                 <button
                   type="button"
                   class="btn-password-visible"
+                  id="passwordVisibleBtn"
                   title="비밀번호 표시/숨김"
                   aria-label="비밀번호 표시/숨김"
                   aria-pressed="false"
