+++ script/auth/signup.js
@@ -4,9 +4,9 @@ import {
   validatePassword,
   validatePasswordCheck,
   validateNickname,
-} from "./util/validators.js";
-import { updateValidationUI } from "./util/updateValidationUI.js";
-import togglePasswordHandler from "./util/togglePassword.js";
+} from "../util/validators.js";
+import { updateValidationUI } from "../util/updateValidationUI.js";
+import togglePasswordVisible from "./togglePasswordVisible.js";
 
 const form = document.querySelector("#signupForm");
 const emailInput = document.querySelector("#userEmail");
@@ -34,6 +34,8 @@ const inputValidatorMap = {
 const FORM_SUBMIT_BUTTON = signupBtn;
 
 const REDIRECT_TARGET = "/login.html";
+
+const PASSWORD_TOGGLE_BUTTONS = form.querySelectorAll(".btn-password-visible");
 
 // 각 키의 유효성 검사값 초기화: [input.id, false]
 const validatorKey = Object.keys(inputValidatorMap);
@@ -91,7 +93,7 @@ function init() {
   delegateFormValidation();
 
   /* 비밀번호 토글 */
-  togglePasswordHandler(form);
+  togglePasswordVisible(PASSWORD_TOGGLE_BUTTONS);
 
   /* 첫번째 input focus 처리 */
   focusFirstInput();
