+++ signup.html
@@ -11,7 +11,7 @@
       href="images/favicon-16x16.png"
     />
     <link rel="stylesheet" href="css/pages/auth.css" />
-    <script src="script/signup.js" type="module" defer></script>
+    <script src="script/auth/signup.js" type="module" defer></script>
   </head>
   <body>
     <main class="signup auth">
@@ -68,6 +68,7 @@
                 <button
                   type="button"
                   class="btn-password-visible off"
+                  id="passwordVisibleBtn"
                   title="비밀번호 표시/숨김"
                   aria-label="비밀번호 표시/숨김"
                   aria-pressed="false"
@@ -92,6 +93,7 @@
                 <button
                   type="button"
                   class="btn-password-visible"
+                  id="passwordChkVisibleBtn"
                   title="비밀번호 표시/숨김"
                   aria-label="비밀번호 표시/숨김"
                   aria-pressed="false"
